Refactor: 관련된 뷰들 하나의 파일로 통합함
