--- VendingMachineApp/VendingMachineApp/Views/BeverageNumberLabel.swift
@@ -1,21 +1,0 @@
-//
-//  BeverageNumberLabel.swift
-//  VendingMachineApp
-//
-//  Created by kimdo2297 on 2020/03/15.
-//  Copyright © 2020 Jason. All rights reserved.
-//
-
-import UIKit
-
-final class BeverageNumberLabel: UILabel {
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-    }
-
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-    }
-    
-}
