+++ VendingMachineApp/VendingMachineApp/Views/BeverageView.swift
@@ -19,3 +19,16 @@ final class PlusBeverageButton: UIButton {
     }
     
 }
+
+final class BeverageNumberLabel: UILabel {
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+    }
+
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+    }
+    
+}
+
